Disable turbolinks due to performance issues
From the turbolink page:
"When you follow a link, Turbolinks automatically fetches the page,
swaps in its <body>, and merges its <head>,
all without incurring the cost of a full page load."
However, in case the current page has extremely large <body> content,
such at the third series in the Carnegie finding aid, the above
mechanism introduces a quite large lag when trying to get to a new
page via a navigation link in the sidebar. Disabling turbolinks
got rid of the lag.