ACFA-99: Add placeholder specifying required date format in aeon Scheduled Date field
The code uses the date_field_tag helper method, which creates an
<input type="date"> in the displayed html code. In Chrome and Firefox,
this html element will be displayed as a datepicker with built-in validation.
The inputed date is then normalized to the yyyy-mm-dd format before being
sent to the server.
Safari and Internet Exploer do not support <input type="date">, and
therefore there is no way to enforce the expected date format. However,
adding a placeholder with the correct format will alert users to the
format to use.