ACFA-250: Configurable components of aeon views
Due to Covid-19, restrictions have been imposed on requests for
boxes. Config settings have been added to display notices and
disable functionality, including login into aeon.
A locale file is now used to store text strings related to the
notices, as well as existing text strings that need to change
due to the current restrictions. In this commit, the text
strings for existing text will contain the original text,
thus creating a record of the original text. A subsequent
commit will be used to update the strings with the new text
describing the restrictions.