ACFA-286: Add aeon requests functionality to all repos
The funcitonality is enable or disabled via the requestable_via_aeon
field in repositories.yml config file. The aeon_site_code field specifies
the site code sent to aeon for each repo. If the checkbox_per_unittitle
field is set to true, each individual item will be selectable, and the
unittitle will be sent in the request. If it is set to false, only
the top level container is selectable, and the unittitle is not sent
since the whole top-level container will be retrieved for the patron.
Also, remove references to RBML in aeon forms