ACFA-376: Selected boxes drepped in view-all
Some of the selected boxes in view-all (all series shown on one page)
do not make it to the resulting aeon request.
BUG: name identifier for the checkbox did not contain the series
number. For example, first checked box in each series was named
checkbox__1, instead of checkbox_1_1 for the first series,
checkbox_2_1 for the second series, and so on. So for each box with
the same checkbox name, only the last box was recorded and made it
to the resulting aeon request.
FIX: include the series number in the checkbox name.
NOTE: this bug only affects the view-all view, not the individual
series view.